Started work on a Digital Miner visual render thing. For some reason it's culling all oceans.

+++ src/main/java/mekanism/client/render/MekanismRenderer.java
@@ -123,7 +123,7 @@ public class MekanismRenderer
 
 		public Block baseBlock = Blocks.sand;
 		
-	    public final void setBlockBounds(float xNeg, float yNeg, float zNeg, float xPos, float yPos, float zPos)
+	    public final void setBlockBounds(double xNeg, double yNeg, double zNeg, double xPos, double yPos, double zPos)
 	    {
 	    	minX = xNeg;
 	    	minY = yNeg;
